+++ packages/jaspr/lib/src/foundation/styles/properties/text.dart
@@ -1,3 +1,4 @@
+import 'angle.dart';
 import 'color.dart';
 import 'unit.dart';
 
@@ -47,8 +48,18 @@ class _ListFontFamily implements FontFamily {
   final List<FontFamily> families;
   const _ListFontFamily(this.families);
 
+  bool _validateFamilies() {
+    if (families.isEmpty) {
+      throw 'FontFamily.list cannot be empty';
+    }
+    return true;
+  }
+
   @override
-  String get value => families.map((f) => f.value).join(', ');
+  String get value {
+    assert(_validateFamilies());
+    return families.map((f) => f.value).join(', ');
+  }
 }
 
 class _VariableFontFamily implements FontFamily {
@@ -113,8 +124,7 @@ class FontStyle {
   static const FontStyle italic = FontStyle._('italic');
   static const FontStyle oblique = FontStyle._('oblique');
 
-  // TODO Angle class
-  const factory FontStyle.obliqueAngle(double degrees) = _ObliqueAngleFontStyle;
+  const factory FontStyle.obliqueAngle(Angle angle) = _ObliqueAngleFontStyle;
 
   static const FontStyle inherit = FontStyle._('inherit');
   static const FontStyle initial = FontStyle._('initial');
@@ -124,12 +134,12 @@ class FontStyle {
 }
 
 class _ObliqueAngleFontStyle implements FontStyle {
-  final double angle;
+  final Angle angle;
 
   const _ObliqueAngleFontStyle(this.angle);
 
   @override
-  String get value => 'oblique ${angle}deg';
+  String get value => 'oblique ${angle.value}';
 }
 
 enum TextTransform {
@@ -213,8 +223,18 @@ class _MultiTextDecorationLine implements TextDecorationLine {
 
   const _MultiTextDecorationLine(this.lines);
 
+  bool _validateLines() {
+    if (lines.isEmpty) {
+      throw 'TextDecorationLine.multi cannot be empty';
+    }
+    return true;
+  }
+
   @override
-  String get value => lines.map((l) => l.value).join(' ');
+  String get value {
+    assert(_validateLines());
+    return lines.map((l) => l.value).join(' ');
+  }
 }
 
 enum TextDecorationStyle {
@@ -288,7 +308,11 @@ class _TextDecoration implements TextDecoration {
   final TextDecorationStyle? style;
   final TextDecorationThickness? thickness;
 
-  const _TextDecoration({this.line, this.color, this.style, this.thickness});
+  const _TextDecoration({this.line, this.color, this.style, this.thickness})
+    : assert(
+        line != null || style != null || color != null || thickness != null,
+        'At least one of line, style or color must not be null. For no text decoration, use TextDecoration.none',
+      );
 
   @override
   String get value => [
@@ -299,15 +323,26 @@ class _TextDecoration implements TextDecoration {
   ].join(' ');
 }
 
-abstract class TextShadow {
-  const factory TextShadow({required Unit offsetX, required Unit offsetY, Unit? blur, Color? color}) = _TextShadow;
+class TextShadow {
+  const TextShadow._(this.value);
 
+  static const none = TextShadow._('none');
+  static const initial = TextShadow._('initial');
+  static const inherit = TextShadow._('inherit');
+  static const revert = TextShadow._('revert');
+  static const revertLayer = TextShadow._('revert-layer');
+  static const unset = TextShadow._('unset');
+
+  const factory TextShadow({required Unit offsetX, required Unit offsetY, Unit? blur, Color? color}) = _TextShadow;
   const factory TextShadow.combine(List<TextShadow> shadows) = _CombineTextShadow;
 
-  String get value;
+  /// The css value
+  final String value;
 }
 
-class _TextShadow implements TextShadow {
+abstract class _ListableTextShadow implements TextShadow {}
+
+class _TextShadow implements _ListableTextShadow {
   const _TextShadow({required this.offsetX, required this.offsetY, this.blur, this.color});
 
   final Unit offsetX;
@@ -316,17 +351,38 @@ class _TextShadow implements TextShadow {
   final Color? color;
 
   @override
-  String get value =>
-      [offsetX.value, offsetY.value, if (blur != null) blur!.value, if (color != null) color!.value].join(' ');
+  String get value => [
+    offsetX.value,
+    offsetY.value,
+    if (blur != null) blur!.value,
+    if (color != null) color!.value,
+  ].join(' ');
 }
 
-class _CombineTextShadow implements TextShadow {
+class _CombineTextShadow implements _ListableTextShadow {
   const _CombineTextShadow(this.shadows);
 
   final List<TextShadow> shadows;
 
+  bool _validateShadows() {
+    if (shadows.isEmpty) {
+      throw 'TextShadow.combine cannot be empty. For no text shadow, use TextShadow.none';
+    }
+
+    for (final shadow in shadows) {
+      if (shadow is! _ListableTextShadow) {
+        throw 'Cannot use ${shadow.value} as a text shadow list item, only standalone use supported';
+      }
+    }
+
+    return true;
+  }
+
   @override
-  String get value => shadows.map((s) => s.value).join(', ');
+  String get value {
+    assert(_validateShadows());
+    return shadows.map((s) => s.value).join(', ');
+  }
 }
 
 enum TextOverflow {