+++ packages/jaspr/lib/src/foundation/styles/properties/transition.dart
@@ -1,11 +1,21 @@
-abstract class Transition {
+class Transition {
+  const Transition._(this.value);
+
+  static const inherit = Transition._('inherit');
+  static const initial = Transition._('initial');
+  static const revert = Transition._('revert');
+  static const revertLayer = Transition._('revert-layer');
+  static const unset = Transition._('unset');
+
   const factory Transition(String property, {required Duration duration, Curve? curve, Duration? delay}) = _Transition;
   const factory Transition.combine(List<Transition> transitions) = _CombineTransition;
 
-  String get value;
+  final String value;
 }
 
-class _Transition implements Transition {
+abstract class _ListableTransition implements Transition {}
+
+class _Transition implements _ListableTransition {
   const _Transition(this.property, {required this.duration, this.curve, this.delay});
 
   final String property;
@@ -22,13 +32,30 @@ class _Transition implements Transition {
   ].join(' ');
 }
 
-class _CombineTransition implements Transition {
+class _CombineTransition implements _ListableTransition {
   const _CombineTransition(this.transitions);
 
   final List<Transition> transitions;
 
+  bool _transitionsListable() {
+    if (transitions.isEmpty) {
+      throw 'Transition.combine cannot be empty.';
+    }
+
+    for (final transition in transitions) {
+      if (transition is! _ListableTransition) {
+        throw 'Cannot use ${transition.value} as a transition list item, only standalone use supported.';
+      }
+    }
+
+    return true;
+  }
+
   @override
-  String get value => transitions.map((t) => t.value).join(', ');
+  String get value {
+    assert(_transitionsListable());
+    return transitions.map((t) => t.value).join(', ');
+  }
 }
 
 class Curve {